isolate changes to request and context per handler
I moved the code from a single location into each handler so its clear
which handler requires the modifications to the context.  This should
allow for others to add their new handlers in a similar manner without
requiring modifications to a single central file.

There is a disadvantage with this approach related to documenting which
methods are available in the context.  I need to think further about if
I want to do this or just leave it the way it was.